remove a bunch of empty lines, clean up indentation, rename index.js to api.js
