+++ client/components/Recipe.js
@@ -172,6 +172,7 @@ const mapDispatchToProps = (dispatch) => {
   }
 
 const mapStateToProps = (state) => {
+  console.log(state.recipe);
   return {
     data: state.recipe
   }
