+++ public/stylesheets/style.css
@@ -7,59 +7,54 @@ a {
   color: #00B7FF;
 }
 
-.jumbotron{
+.jumbotron {
   background-color: white;
   border: none;
 }
-.row.content{
 
-margin:10px;
-padding: 5px;
-
+.row.content {
+  margin:10px;
+  padding: 5px;
 }
 
-li{
+li {
   font-size: 20px;
   margin-bottom:10px;
-
 }
 
-.nav{
+.nav {
   margin: 10px;
 }
 /* adds padding to navigation links on header*/
-.navli{
+.navli {
   padding:5px;
 }
 
-.well{
+.well {
   border: none;
   background-color: white;
 }
 
-.thumbnail{
+.thumbnail {
   border: none;
 }
 
-.accordion{
-
+.accordion {
  background-color: crimson;
  color: white;
 }
-.collapse{
+
+.collapse {
  margin: 10px;
  padding: 5px;
 }
-.collapse h5{
 
+h5 {
  color:black;
  font-weight: 400;
 }
 
-.left-text{
-
+.left-text {
  font-weight: bold;
  padding-bottom: 15px;
-
-
 }
